+++ lib/src/howtoplayscreen.dart
@@ -7,11 +7,16 @@ class HowToPlayScreen extends InfoScreen {
 
   @override
   void initialize() {
+    add('Soundtrack:', fgColor: Color.orange);
+    add('The Speed Vector Unofficial Soundtrack can be found on Spotify. It is not required for playing the game, but is highly recommended!');
+    add('https://open.spotify.com/playlist/', fgColor: Color.orange);
+    add('   3u5g9wybWolaozoC5SAedt', fgColor: Color.orange);
+    add(' ');
     add('The basic idea of SPEED VECTOR is simple:');
     add('You drive your car along a never-ending, procedurally generated stretch of road.');
-    add('You must stay on the road, and try to not get too banged up by the other motorists in the process.');
+    add('You must stay on the road and try to not get too banged up by the other motorists in the process.');
     add(' ');
-    add('Controlling your car is all about momentum. The faster you go, the more difficult it will be to stay on the road.');
+    add('Controlling your car is all about momentum. The faster you go, the more difficult it will be to control your vehicle.');
     add(' ');
     add('More info for the mathematically inclined:', fgColor: Color.darkGold);
     add('<<< DON\'T PANIC! If this makes no sense to you, just skip it. You don\'t need to fully understand it to play the game. >>>',
@@ -23,7 +28,7 @@ class HowToPlayScreen extends InfoScreen {
     add(' ');
     add('You change direction by picking one of the (usually nine) possible choices using your cursor: +');
     add(' ');
-    add('Whatever point you choose, that is where you will (generally) end up after your move.');
+    add('Whatever point you choose, that is where you will (generally) end up next turn.');
     add('Exceptions exist. For instance, you can be knocked slightly off course by bumping into other cars.',
         fgColor: Color.gray);
     add(' ');
@@ -33,7 +38,7 @@ class HowToPlayScreen extends InfoScreen {
     add(' ');
     add('                        τ', fgColor: Color.purple);
     add(' ');
-    add('Pick one of the three top directions and you speed up. The middle three means you stay at your current speed. The bottom three lowers your speed, allowing you to take tighter turns.');
+    add('Pick one of the three top directions and you speed up. The middle three means you stay at your current speed. The bottom three lowers your speed, allowing you to make tighter turns.');
     add(' ');
     add('But watch out! If you go too slow, or too fast, you risk moving off screen and being eliminated!',
         fgColor: Color.red);
@@ -45,11 +50,11 @@ class HowToPlayScreen extends InfoScreen {
     add(' ');
     add('When you\'re ready for even more speed, you can use the numpad keys [1-9] or the mouse to instantly select a direction and move, without having to manually confirm your choice.');
     add(' ');
-    add('After all cars have moved, the road "rolls" down a number of steps, depending on your current speed. For every piece of road that disappears from the bottom of the screen, a new piece of fresh road appears at the top.');
+    add('At the end of each turn, after all cars have moved, the road "rolls" down a number of steps, depending on your current speed. For every piece of road that disappears from the bottom of the screen, a new piece of fresh road appears at the top.');
     add(' ');
     add('That\'s basically all you need to know to get started. Don\'t worry if you don\'t understand exactly everything now.');
     add('Once you actually start playing, you will pick it up in no time!');
     add(' ');
-    add('Now go and burn some rubber!!!', fgColor: Color.orange);
+    add('Now go burn some rubber!!!', fgColor: Color.orange);
   }
 }